cliconfig: Do not rely on platform dependent code
The `github.com/docker/docker/pkg/system` import was only being used to
call to os.MkdirAll so it seems unnecessary to pull in all of system/*
when this is all that is necessary.

Additionally, this was causing problems with attempting to vendor
cliconfig using Godeps, as it would only vendor the necessary files
based on the platform it's running on, meaning the `meminfo_*.go` files
were being missed, also causing `github.com/docker/docker/pkg/units` to
be missed.

Signed-off-by: Chance Zibolski <chance@coreos.com>